--- cli/src/views/App/Contract/modules/ContractHeader/index.js
@@ -1,2 +1,0 @@
-import ContractHeader from './components/ContractHeader'
-export default ContractHeader
