+++ cli/src/views/App/Contract/modules/ContractSidebar/components/ContractSidebar.jsx
@@ -1,11 +1,24 @@
 import React, { useState, useEffect, useCallback } from 'react'
 import PropTypes from 'prop-types'
-import { Row, Col, Card } from 'antd'
-import ContractNavs from '../modules/ContractNavs'
+import { Card } from 'antd'
+import { FILTERS } from '../modules/const'
 import { apiRoom } from 'constants/api'
+import _ from 'lodash'
 
-export default function ContractSidebar({ children, loading, ...props }) {
+export default function ContractSidebar({ children, loading, onSelectRoom }) {
   const [rooms, setRooms] = useState([])
+  const [roomId, setRoomId] = useState(null)
+
+  const selectRoom = useCallback(roomId => {
+    setRoomId(roomId)
+    onSelectRoom(roomId)
+  }, [onSelectRoom])
+
+  useEffect(() => {
+    // Auto select the first room after accessing to the page
+    const noSelectedRoom = rooms.length > 0 && !roomId
+    noSelectedRoom && selectRoom(rooms[0]._id)
+  }, [roomId, rooms, selectRoom])
 
   const init = useCallback(async() => {
     let rooms = await apiRoom.get()
@@ -15,30 +28,7 @@ export default function ContractSidebar({ children, loading, ...props }) {
 
   useEffect(() => { init() }, [init])
 
-  // return (
-  //   <>
-  //     <div className='gx-d-block gx-d-lg-none'>
-  //       <Drawer
-  //         placement='left'
-  //         closable={false}
-  //         visible={false}
-  //         onClose={() => {}}
-  //       >
-  //         <ContractNavs rooms={rooms} {...props}>
-  //           { children }
-  //         </ContractNavs>
-  //       </Drawer>
-  //     </div>
-  //     <div className='gx-module-sidenav gx-d-none gx-d-lg-flex'>
-  //       <ContractNavs rooms={rooms} {...props}>
-  //         { children }
-  //       </ContractNavs>
-  //     </div>
-  //   </>
-  // )
-
   return (
-    // <Card className='gx-card contract-sidebar' title='Available Contracts For:'>
     <Card className='gx-card contract-sidebar'>
       <div className='gx-module-side-header'>
         <div className='gx-module-logo'>
@@ -47,14 +37,42 @@ export default function ContractSidebar({ children, loading, ...props }) {
         </div>
       </div>
       {children}
-      <ContractNavs rooms={rooms} {...props}/>
+      <ul className='gx-module-nav'>
+        <div className='section-wrapper'>
+          <div>Active Contracts</div>
+          {
+            _.map(rooms, r => (
+              <li onClick={selectRoom.bind(this, r._id)} key={r._id}>
+                <span className={roomId === r._id ? 'gx-link active' : 'gx-link'}>
+                  <i className='icon icon-schedule'/>
+                  <span>{r.name}</span>
+                </span>
+              </li>
+            ))
+          }
+        </div>
+        <div className='section-wrapper'>
+          <div>Status Filters</div>
+          {
+            _.map(FILTERS, (f, i) => (
+              <li onClick={() => {}} key={i}>
+                <span className='gx-link'>
+                  <i className='icon icon-chart-pie'/>
+                  <span>{f}</span>
+                </span>
+              </li>
+            ))
+          }
+        </div>
+      </ul>
     </Card>
   )
 }
 
 ContractSidebar.propTypes = {
   children: PropTypes.any,
-  loading: PropTypes.func
+  loading: PropTypes.func,
+  onSelectRoom: PropTypes.func
 }
 
 ContractSidebar.defaultProps = {
