--- cli/src/views/App/Contract/modules/ContractSidebar/modules/ContractNavs/index.js
@@ -1,2 +1,0 @@
-import ContractNavs from './components/ContractNavs'
-export default ContractNavs
