+++ cli/src/views/App/Contract/modules/ContractSidebar/modules/const.js
@@ -1,0 +1,1 @@
+export const FILTERS = ['Pending', 'Processing', 'Complete']
